Improve the binary format
With this change each work item section header describes the format of each contained work item. Every attribute is encoded on optimal number of bytes, as a difference from the base value. This imposes a significant performance hit on both serializing and deserializing the file, yet it greatly reduces its size. This tradeoff is acceptable mainly due the fact, that this algorithm was planned to be implemented from the very beginning.